Fix signed <-> unsigned comparisons
This mostly means casting index-loop sentinels to the index type. For comparisons against the `FramePointers` `SampleCount`, we can just make the latter an `int` instead, which is sufficient to hold all thoretically-valid values.

+++ src/audio_backend.cpp
@@ -100,7 +100,7 @@ void RealTimeAudioThread::AdvanceFrames(FramePointers& Frame)
 
     FramePressureCount = std::max(FramePressureIndex, FramePressureCount);
     FramePressureIndex %= FramePressure.size();
-    if (Program && FramePressureCount == FramePressure.size())
+    if (Program && FramePressureCount == static_cast<int>(FramePressure.size()))
     {
         TemporalPressure = FramePressure[0];
         for (int Index = 1; Index < FramePressureCount; ++Index)
